requirements.txt
pypdf==5.4.0
nltk
openai
python-dotenv
